requirements.txt

numpy==1.24.4
openai==1.16.2
pandas==2.0.3
pathlib==1.0.1
pdfminer.six==20231228
pillow==10.3.0
requests==2.31.0
six==1.16.0
tqdm==4.66.2